Move message factories to their own package
