+++ toolkit/src/main/java/com/onelogin/saml2/factory/SamlReceivedMessageFactory.java
@@ -1,4 +1,4 @@
-package com.onelogin.saml2;
+package com.onelogin.saml2.factory;
 
 import com.onelogin.saml2.http.HttpRequest;
 import com.onelogin.saml2.settings.Saml2Settings;
